refactor(layout): flatten commands/ into cli/commands/
