+++ src/cli/commands/__tests__/doctor.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi } from 'vitest';
-import { doctor, formatDoctor } from '../index.js';
+import { doctor, formatDoctor } from '../doctor.js';
 import type { ProfileRepository } from '../../../repository/profileRepository.js';
 import type { Profile } from '../../../utils/types/index.js';
 import type { AppContext } from '../../../runtime/appContext.js';
