+++ src/cli/commands/__tests__/env.test.ts
@@ -2,7 +2,7 @@ import fs from 'node:fs/promises';
 import os from 'node:os';
 import path from 'node:path';
 import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
-import { writeWolfBlock, envSetOne } from './index.js';
+import { writeWolfBlock, envSetOne } from '../env.js';
 
 let tmpDir: string;
 let rcFile: string;
