+++ src/cli/commands/__tests__/job-list.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
-import { jobList, formatJobList, runJobListCli } from '../list.js';
+import { jobList, formatJobList, runJobListCli } from '../job/list.js';
 import type { AppContext } from '../../../runtime/appContext.js';
 import type { Job, JobQuery } from '../../../utils/types/job.js';
 import type { Company } from '../../../utils/types/company.js';
