+++ src/cli/commands/add.ts
