+++ src/cli/commands/doctor.ts
