+++ src/cli/commands/hunt.ts
