+++ src/cli/commands/profile.ts
