+++ src/cli/commands/score.ts
