+++ src/cli/commands/status.ts
